cli fixed
