+++ digital_twin_distiller/__init__.py
@@ -18,6 +18,4 @@ from digital_twin_distiller.platforms import *
 from digital_twin_distiller.snapshot import Snapshot
 from digital_twin_distiller.model import BaseModel
 
-from .__main__ import new
-
 __all__ = ['new']
