+++ tests/test_encapsulator.py
@@ -1,17 +1,14 @@
-import os
 import unittest
-from multiprocessing import Process
 from pathlib import Path
-from time import sleep
 
 import requests
 from fastapi.testclient import TestClient
 
-from digital_twin_distiller.__main__ import new
 from digital_twin_distiller.encapsulator import Encapsulator
 from digital_twin_distiller.modelpaths import ModelDir
-from digital_twin_distiller.simulationproject import SimulationProject, sim
+from digital_twin_distiller.simulationproject import SimulationProject
 from digital_twin_distiller.utils import purge_dir
+from digital_twin_distiller.cli import new
 
 CURRENT = Path(__file__).parent
 MODELNAME = "TestServer"
